own player class
